+++ pages/_app.page.tsx
@@ -2,7 +2,7 @@ import type { AppProps } from 'next/app';
 import styled from '@emotion/styled';
 import { Global } from '@emotion/react';
 import { globalStyles } from 'src/shared';
-import NavBar from 'src/components/navigation/NavBar';
+import NavBar from 'src/components/navigation/NavBar/NavBar';
 import { useRouter } from 'next/router';
 import { RecoilRoot } from 'recoil';
 import Head from 'src/components/Head';
