+++ src/components/navigation/NavBar/NavBar.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import Link from 'next/link';
 import { useRouter } from 'next/router';
 import styled from '@emotion/styled';
-import { Menus } from './Menus';
-import WritePageMenus from './WritePageMenus';
+import { Menus } from '../Menus';
+import WritePageMenus from '../WritePageMenus';
 
 const NavBar = () => {
   const router = useRouter();
